Assign fixed values to your chart width and height
This piece of VBA code allows you to give fixed dimensions to your graphs in Excel and avoids them to change into what Excel believes is the optimal size. Be sure to have the graph you want to fix the size with as active (right click on the graph) and run the code by assigning the macro to this graph.